+++ fixtures/gists-app/data/routes/gists.js
@@ -1,8 +1,17 @@
+const { json } = require("@remix-run/loader");
+
 module.exports = () => {
-  return {
-    users: [
-      { id: "ryanflorence", name: "Ryan Florence" },
-      { id: "mjackson", name: "Michael Jackson" }
-    ]
-  };
+  return json(
+    {
+      users: [
+        { id: "ryanflorence", name: "Ryan Florence" },
+        { id: "mjackson", name: "Michael Jackson" }
+      ]
+    },
+    {
+      headers: {
+        "Cache-Control": "public, max-age=60"
+      }
+    }
+  );
 };