+++ packages/react/markup.ts
@@ -2,6 +2,6 @@ export interface SafeHtml {
   __html: string;
 }
 
-export default function createHtml(html: string): SafeHtml {
+export function createHtml(html: string): SafeHtml {
   return { __html: html };
 }